Extract image gallery related functions into a separate class
