+++ src/components/Sidebar.js
@@ -1,5 +1,4 @@
-import ShadowDOMContainer from "./ShadowDOMContainer";
-import { promisifyImageLoad, updateLabel } from "../utility/helper";
+import ImageGallery from "../components/ImageGallery";
 
 const SidebarStyles = {
   height: "100%",
@@ -15,42 +14,28 @@ const SidebarStyles = {
  * Sidebar component with built-in logic and styles
  */
 export default class Sidebar {
-  constructor() {
-    this.shadowDOMContainer = new ShadowDOMContainer();
+  constructor(shadowDOMContainer) {
+    this.shadowDOMContainer = shadowDOMContainer;
+    this.imageGallery = new ImageGallery(shadowDOMContainer);
     this.isShown = false;
-    this.inActiveElement = this.addInActiveSurfaceElement();
-
-    // Defined here to grab the correct invocation context (this) when invoked from inside the proxy
-    const updateSidebar = (size) => {
-      this.shadowDOMContainer.container.querySelector("#send-btn").disabled = !(
-        size > 0
-      );
-      updateLabel(this.shadowDOMContainer.container, "selected-images", size);
-    };
-
-    this.selectedImages = new Proxy(new Map(), {
-      get(target, property) {
-        if (property === "size") {
-          return target.size;
-        }
-
-        if (property === "set" || property === "delete") {
-          const operator = { set: 1, delete: -1 };
-          const size = target.size + operator[property];
-          updateSidebar(size);
-        }
-
-        let value = Reflect.get(...arguments);
-        return typeof value === "function" ? value.bind(target) : value;
-      },
-    });
+    this.inActiveElement = this.addInActiveBgElement();
   }
 
-  // To shadow the background when the drawer is open
-  addInActiveSurfaceElement() {
-    const inActiveSurfaceElement = document.createElement("div");
+  /**
+   * Initialize the markup, styles and other child components
+   */
+  async init(results) {
+    await this.shadowDOMContainer.init(SidebarStyles);
+    this.imageGallery.init(results);
+  }
 
-    inActiveSurfaceElement.css({
+  /**
+   * Dark background shade overlayed on the web page to add more focus to the sidebar
+   */
+  addInActiveBgElement() {
+    const inActiveBgElement = document.createElement("div");
+
+    inActiveBgElement.css({
       backgroundColor: "rgba(0, 0, 0)",
       bottom: "0",
       left: "0",
@@ -63,75 +48,14 @@ export default class Sidebar {
     });
 
     // Close the sidebar if the user clicks outside the sidebar's scope
-    inActiveSurfaceElement.addEventListener("click", () => {
+    inActiveBgElement.addEventListener("click", () => {
       this.closeSidebar();
     });
 
-    document.body.prepend(inActiveSurfaceElement);
+    document.body.prepend(inActiveBgElement);
 
-    return inActiveSurfaceElement;
+    return inActiveBgElement;
   }
-
-  async init() {
-    return await this.shadowDOMContainer.init(SidebarStyles);
-  }
-
-  loadImagesOntoGallery(images) {
-    const [
-      showcase,
-      loader,
-    ] = this.shadowDOMContainer.container.querySelectorAll(
-      ".gallery__showcase, .gallery__loader"
-    );
-    let pImages = [];
-
-    // Make each image unique, add listeners to be used to later to shake off broken images
-    for (let index = 0; index < images.length; index++) {
-      const image = images[index];
-      const tileImage = document.createElement("img");
-      tileImage.setAttribute("src", image.src);
-      tileImage.setAttribute("alt", "");
-      tileImage.dataset.index = index;
-      tileImage.classList.add("tile");
-      tileImage.addEventListener("click", this.onImageSelected.bind(this));
-      // Collect list of broken images / successfully loaded ones
-      pImages.push(promisifyImageLoad(tileImage));
-    }
-
-    // Inject the loaded images onto the gallery, dismiss broken images
-    Promise.allSettled(pImages)
-      .then((results) => results.filter(({ status }) => status === "fulfilled"))
-      .then((loadedImages) => {
-        loader.parentNode.removeChild(loader); // Loader suicides
-
-        for (const { value: image } of loadedImages) {
-          const layer = document.createElement("div"); // Add inner div as a layer over the image
-          layer.classList.add("tile__layer");
-          layer.appendChild(image);
-          showcase.appendChild(layer); // Tile added to gallery
-        }
-
-        updateLabel(
-          this.shadowDOMContainer.container,
-          "total-images",
-          loadedImages.length
-        );
-      });
-  }
-
-  onImageSelected(e) {
-    const { index } = e.target.dataset;
-
-    if (e.target.parentElement.classList.contains("tile--selected")) {
-      this.selectedImages.delete(index);
-    } else {
-      this.selectedImages.set(index, e.target);
-    }
-
-    e.target.parentElement.classList.toggle("tile--selected");
-  }
-
-  // Toggle opening / closing sidebar
 
   openSidebar() {
     this.isShown = true;
