+++ src/content/content.js
@@ -1,15 +1,16 @@
 import { ImgMetaDataAPI } from "../utility/ImgMetaDataAPI";
 import { inlineCSS } from "../utility/helper";
 import Sidebar from "../components/Sidebar";
+import ShadowDOMContainer from "../components/ShadowDOMContainer";
 
 HTMLDivElement.prototype.css = inlineCSS;
 
 const results = ImgMetaDataAPI.getImgMetaData();
-const sidebar = new Sidebar();
+const shadowDOMContainer = new ShadowDOMContainer();
+const sidebar = new Sidebar(shadowDOMContainer);
 
-sidebar.init().then(() => {
+sidebar.init(results).then(() => {
   chrome.runtime.onMessage.addListener(onMessageListener);
-  sidebar.loadImagesOntoGallery(results);
 });
 
 /**
